extract ProductsLoader abstract class
