+++ lib/products_feature/product.dart
@@ -1,18 +1,3 @@
-class Products {
-  List<Product> data;
-
-  Products({required this.data});
-
-  factory Products.fromJson(Map<String, dynamic> json) {
-    List<Product> products = [];
-    final data = json['data'];
-    data.forEach((value) {
-      products.add(Product.fromJson(value));
-    });
-    return Products(data: products);
-  }
-}
-
 class Product {
   String sku;
   String name;
