+++ lib/products_feature/products_api/api_products_model.dart
@@ -1,0 +1,16 @@
+import 'package:thepos/products_feature/product.dart';
+
+class Products {
+  List<Product> data;
+
+  Products({required this.data});
+
+  factory Products.fromJson(Map<String, dynamic> json) {
+    List<Product> products = [];
+    final data = json['data'];
+    data.forEach((value) {
+      products.add(Product.fromJson(value));
+    });
+    return Products(data: products);
+  }
+}
