+++ lib/products_feature/products_loader.dart
@@ -1,0 +1,5 @@
+import 'package:thepos/products_feature/product.dart';
+
+abstract class ProductsLoader {
+  Future<List<Product>> loadProducts();
+}
